Include Arev license covering imported glyphs
Fixes issue #106

The DejaVu Fonts project imported glyphs from the Arev Fonts project
covering extended character ranges. As these glyphs are also included in
Hack, the appropriate license and credit should accompany the source in
the same way that DejaVu has done. To this end the appropriate license
wording covering that subset of glyphs is included below as found on the
DejaVu project site and the upstream Arev Fonts site.