Finish updating Arch Linux installation instructions
Somebody on the previous PR mentioned that the ttf-hack package moved
from the AUR to the community repository. See:

https://github.com/chrissimpkins/Hack/pull/151#discussion_r48590088

This was correct information and the change in 3e47153 to a different
package manager is also correct, but the package URL and the description
also needs to match. Installing community package cannot be done
"manually" nor does mentioning an "AUR helper" make sense any more.  If
the official core package manager is involved (yeh!) then just let it do
its job.