Redraw `zero` in Regular
The shape of the oval in the zero are merged offset paths with distances equal to the horizontal and vertical stroke width. I've manually merged the coordinates to produce a smooth shape. There is still a discrepancy in the coordinates, because the left and right halves of the original glyph aren't exactly mirrored.