Powerline: Improve rendering at small size/hidpi
These changes to the Powerline separator glyphs is the result of
studying the ones from DejaVu Sans Mono and Fantasque Sans Mono which
appear seamless, without gaps or points extending above and below the
line.  We make the outlines slightly larger than the bounding box to
cause the renderer to overlap the character with the surrounding ones,
eliminating the gaps.  Experimentation with different sizes and shapes
was performed in #234 under several renderers and applications to try to
get the most consistency across platforms.

Closes #33